requirements_advanced.txt
ccxt>=4.1.68
pandas>=2.1.4
numpy>=1.24.3
TA-Lib==0.4.28
python-dotenv==1.0.0
scikit-learn==1.3.2
ta==0.10.2
schedule==1.2.0
websocket-client==1.6.4
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
